add call transcript function to functions.py, add comments planning next step
